Changing restoring dependencies (npm ci)
In the Restoring Dependencies section (line 48) we are telling users to run an npm install. That is slow and unnecessarily so. By adding the continuous integration 'ci' argument to an 'npm ci' the package-lock.json file is used instead of package.json. This removes the need to verify libraries during the install. In my testing this has increased the speed of the step by at least 50%. As this is by far and away the slowest step of the build process I think it's important. I am happy to add more description to the article if necessary.

Reference: https://stackoverflow.com/questions/52499617/what-is-the-difference-between-npm-install-and-npm-ci/53325242#53325242

A replacement image can be found here
https://xomino.files.wordpress.com/2019/06/spfx-restoringdepenciesupdate_mdr.jpg